requirements.txt
Django==3.0.8 
django-jet==1.0.8
Pillow==7.0.0
markdown==3.2
django-filter==2.2.0
psycopg2-binary==2.8.5
python-dotenv==0.14.0